refactor(enrollment): move EnrollmentPolicy into mills
The policy aggregates percentages and decides who may enrol, which is
business logic; pacts is contracts. It sat there only because
Event.enrollment_policy was an ORM method and links should not reach into
mills — so that wrapper goes, and its two callers build the policy directly.

Drops banner_text and allows_anonymous_enrollment: neither had a production
caller, and anonymous access is still combined by its own any() in links
(#712). A value object shipping speculative accessors invites the drift this
work exists to remove.

Also drops the per-request policy memo. It was keyed on a session argument
it ignored, and the call it saves is a DB read — the membership API is
already guarded by allowed_slots and a check interval.
