+++ src/ludamus/mills/enrollment.py
@@ -9,10 +9,13 @@ the ticket API.
 
 from __future__ import annotations
 
+import math
 import secrets
+import sys
+from dataclasses import dataclass
 from datetime import UTC, datetime, timedelta
 from secrets import token_urlsafe
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Protocol
 
 from ludamus.pacts import (
     MembershipAPIError,
@@ -51,6 +54,8 @@ from ludamus.pacts.party import HeldSeatNotification
 from ludamus.specs.enrollment import is_valid_window_period, select_promotable_parties
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
+
     from ludamus.pacts import (
         EnrollmentConfigDTO,
         EnrollmentConfigRepositoryProtocol,
@@ -82,6 +87,78 @@ if TYPE_CHECKING:
     from ludamus.pacts.services import TransactionProtocol
 
 _NAVBAR_NOTIFICATION_LIMIT = 10
+
+
+class EnrollmentWindowLike(Protocol):
+    max_waitlist_sessions: int
+    percentage_slots: int
+    restrict_to_configured_users: bool
+
+
+def _percentage_slots(window: EnrollmentWindowLike) -> int:
+    return window.percentage_slots
+
+
+@dataclass(frozen=True)
+class EnrollmentPolicy:
+    """What one actor may do across the enrollment windows open to them.
+
+    A window grants access for its period, so the windows an actor can use are
+    unioned. Selecting first and aggregating second is what keeps capacity from
+    being drawn from a window the actor is not allowed into.
+    """
+
+    windows: tuple[EnrollmentWindowLike, ...]
+
+    @classmethod
+    def for_actor(
+        cls, windows: Iterable[EnrollmentWindowLike], *, is_configured_user: bool
+    ) -> EnrollmentPolicy:
+        return cls(
+            tuple(
+                window
+                for window in windows
+                if is_configured_user or not window.restrict_to_configured_users
+            )
+        )
+
+    @property
+    def can_enroll(self) -> bool:
+        return bool(self.windows)
+
+    @property
+    def seating_window(self) -> EnrollmentWindowLike | None:
+        if not self.windows:
+            return None
+        return max(self.windows, key=_percentage_slots)
+
+    @property
+    def percentage_slots(self) -> int:
+        return self.seating_window.percentage_slots if self.seating_window else 0
+
+    @property
+    def max_waitlist_sessions(self) -> int:
+        return max((window.max_waitlist_sessions for window in self.windows), default=0)
+
+    @property
+    def restricts_everyone(self) -> bool:
+        return bool(self.windows) and all(
+            window.restrict_to_configured_users for window in self.windows
+        )
+
+    @property
+    def requires_slot_allowance(self) -> bool:
+        return bool(
+            self.seating_window and self.seating_window.restrict_to_configured_users
+        )
+
+    def available_slots(self, *, participants_limit: int, enrolled_count: int) -> int:
+        if not self.windows:
+            return 0
+        if participants_limit == 0:
+            return sys.maxsize
+        effective_limit = math.ceil(participants_limit * self.percentage_slots / 100)
+        return max(0, effective_limit - enrolled_count)
 
 
 def _now() -> datetime:
